+++ noteable/sql/meta_commands.py
@@ -15,11 +15,9 @@ from IPython.display import HTML, display
 from pandas import DataFrame
 
 # Will be good when this layer of code doesn't import anything sqlalchemy
-from sqlalchemy.engine.reflection import Inspector
 from sqlalchemy.exc import NoSuchTableError
-from sqlalchemy.types import TypeEngine
 
-from noteable.sql.connection import Connection, IntrospectableConnection
+from noteable.sql.connection import Connection, InspectorProtocol, IntrospectableConnection
 from noteable.sql.types import (
     CheckConstraintModel,
     ColumnModel,
@@ -120,8 +118,9 @@ class MetaCommand:
         """
         raise NotImplementedError
 
-    def get_inspector(self) -> Inspector:
-        return self.conn._get_inspector()
+    def get_inspector(self) -> InspectorProtocol:
+        """Return a Noteable InspectorProtocol instance from this Connection"""
+        return self.conn.get_inspector()
 
 
 class SchemasCommand(MetaCommand):
@@ -230,7 +229,7 @@ class ViewsCommand(MetaCommand):
 
 
 def relation_names(
-    inspector: Inspector,
+    inspector: InspectorProtocol,
     argument: str,
     include_tables=True,
     include_views=True,
@@ -252,6 +251,7 @@ def relation_names(
         '*foo' -> Find all relations in default schema whose name ends with 'foo.'
         'foo??' -> Find all relations in default schema whose names start with 'foo' and only has two following letters in name.
     """
+
     schema_name_glob, relation_name_glob = parse_schema_and_relation_glob(inspector, argument)
 
     schema_name_filter = convert_relation_glob_to_regex(schema_name_glob)
@@ -267,18 +267,14 @@ def relation_names(
         relation_types: List[str] = []
 
     for schema in schemas:
-        # Some dialects return views as tables (and then also as views), so distict-ify via a set.
-        relations = set()
-        if include_tables:
-            relations.update(inspector.get_table_names(schema))
+        relations: List[str] = []
 
-        view_names = inspector.get_view_names(schema)
+        if include_tables:
+            relations.extend(inspector.get_table_names(schema))
+
         if include_views:
-            relations.update(view_names)
-        else:
-            # Because some dialects may have already included view names in get_table_names(), need
-            # to explicitly remove the definite view names. Thanks, guys.
-            relations.difference_update(view_names)
+            view_names = set(inspector.get_view_names(schema))
+            relations.extend(view_names)
 
         # Filter, sort, append schema, relname and possibly the kind onto respective lists.
         for relname in sorted(r for r in relations if relation_name_filter.match(r)):
@@ -310,7 +306,7 @@ def relation_names(
 
 
 def parse_schema_and_relation_glob(
-    inspector: SchemaStrippingInspector, schema_and_possible_relation: str
+    inspector: InspectorProtocol, schema_and_possible_relation: str
 ) -> Tuple[str, str]:
     """Return tuple of schema name glob, table glob given a single string
     like '*', 'public.*', 'foo???', ...
@@ -320,6 +316,9 @@ def parse_schema_and_relation_glob(
 
     Expects to be driven with at least a single character string.
     """
+
+    # TODO: Because some dialects support three-tier nomenclatures (database.schema.relation), this
+    # might well need to move into per-Connection and/or InspectorProtocol implementation. But not today.
 
     if schema_and_possible_relation == '.':
         # Degenerate value. Treat like they meant wildcard
@@ -413,17 +412,10 @@ class SingleRelationCommand(MetaCommand):
 
         for col in column_dicts:
             names.append(col['name'])
-
-            # Convert the possibly db-centric TypeEngine instance to a sqla-generic type string
-            type_name = determine_column_type_name(col['type'])
-
-            types.append(type_name)
+            types.append(col['type'])
             nullables.append(col['nullable'])
             defaults.append(col['default'])
-            if 'comment' in col:
-                # Dialect may not return this attribute at all. If supported by dialect,
-                # but not present on this column, should be the empty string.
-                comments.append(col['comment'])
+            comments.append(col['comment'])
 
         # Assemble dataframe out of data, conditionally skipping columns if inappropriate
         # or zero-value.
@@ -509,16 +501,6 @@ class IntrospectAndStoreDatabaseCommand(MetaCommand):
 
     include_in_help = False
 
-    # How many threads to use to introspect individual tables or views.
-    MAX_INTROSPECTION_THREADS = 10
-
-    # Schemas to never introspect into. Will need to augment / research for each
-    # new datasource type supported.
-    # Clickhouse spells it as 'INFORMATION_SCHEMA' (all caps), rest as 'information_schema' (all lowercase)
-    # Clickhouse also has a 'system' schema, but we don't want to avoid introspecting that
-    # at the risk of excluding user-created 'system' schemas in other datasources.
-    AVOID_SCHEMAS = {'information_schema', 'INFORMATION_SCHEMA', 'pg_catalog', 'crdb_internal'}
-
     def run(self, invoked_as: str, args: List[str]) -> None:
         """Drive introspecting whole database, POSTing results back to Gate for storage.
 
@@ -571,9 +553,7 @@ class IntrospectAndStoreDatabaseCommand(MetaCommand):
 
             # Introspect each relation concurrently.
             # TODO: Take minimum concurrency as a param?
-            with ThreadPoolExecutor(
-                max_workers=self.get_max_threadpool_workers(inspector)
-            ) as executor:
+            with ThreadPoolExecutor(max_workers=inspector.max_concurrency) as executor:
                 future_to_relation = {
                     executor.submit(
                         self.fully_introspect, inspector, schema_name, relation_name, kind
@@ -597,39 +577,16 @@ class IntrospectAndStoreDatabaseCommand(MetaCommand):
     ###
 
     @classmethod
-    def get_max_threadpool_workers(cls, inspector: SchemaStrippingInspector) -> int:
-        """Determine max concurrency for introspecting this sort of database.
-
-        BigQuery, in particular, seems to be not totally threadsafe, ENG-5808
-        """
-
-        if inspector.engine.driver == 'bigquery':
-            return 1
-        else:
-            return cls.MAX_INTROSPECTION_THREADS
-
-    @classmethod
-    def all_table_and_views(cls, inspector) -> List[Tuple[str, str, str]]:
+    def all_table_and_views(cls, inspector: InspectorProtocol) -> List[Tuple[str, str, str]]:
         """Returns list of (schema name, relation name, table-or-view) tuples"""
 
         results = []
 
-        default_schema = inspector.default_schema_name
-        all_schemas = set(inspector.get_schema_names())
-        all_schemas.difference_update(cls.AVOID_SCHEMAS)
-        if default_schema and default_schema not in all_schemas:
-            all_schemas.add(default_schema)
+        all_schemas = inspector.get_schema_names()
 
         for schema_name in sorted(all_schemas):
-            table_names = set(inspector.get_table_names(schema_name))
+            table_names = inspector.get_table_names(schema_name)
             view_names = inspector.get_view_names(schema_name)
-
-            # Some dialects (lookin' at you, cockroach) return view names as both view
-            # names and table names. Sigh. We'd like to only return the counts of
-            # the definite tables, though, so ...
-            if view_names:
-                # Remove any view names from our pristeen list of table names.
-                table_names.difference_update(view_names)
 
             for table_name in table_names:
                 results.append((schema_name, table_name, 'table'))
@@ -641,7 +598,7 @@ class IntrospectAndStoreDatabaseCommand(MetaCommand):
 
     @classmethod
     def fully_introspect(
-        cls, inspector: SchemaStrippingInspector, schema_name: str, relation_name: str, kind: str
+        cls, inspector: InspectorProtocol, schema_name: str, relation_name: str, kind: str
     ) -> RelationStructureDescription:
         """Drive introspecting into this single relation, making all the necessary Introspector API
         calls to learn all of the relation's sub-structures.
@@ -695,7 +652,7 @@ class IntrospectAndStoreDatabaseCommand(MetaCommand):
 
     @classmethod
     def introspect_foreign_keys(
-        cls, inspector: SchemaStrippingInspector, schema_name: str, relation_name: str
+        cls, inspector: InspectorProtocol, schema_name: str, relation_name: str
     ) -> List[ForeignKeysModel]:
         """Introspect all foreign keys for a table, describing the results as a List[ForeignKeysModel]"""
 
@@ -727,7 +684,7 @@ class IntrospectAndStoreDatabaseCommand(MetaCommand):
 
     @classmethod
     def introspect_check_constraints(
-        cls, inspector, schema_name, relation_name
+        cls, inspector: InspectorProtocol, schema_name: str, relation_name: str
     ) -> List[CheckConstraintModel]:
         """Introspect all check constraints for a table, describing the results as a List[CheckConstraintModel]"""
 
@@ -746,7 +703,7 @@ class IntrospectAndStoreDatabaseCommand(MetaCommand):
 
     @classmethod
     def introspect_unique_constraints(
-        cls, inspector, schema_name, relation_name
+        cls, inspector: InspectorProtocol, schema_name: str, relation_name: str
     ) -> List[UniqueConstraintModel]:
         """Introspect all unique constraints for a table, describing the results as a List[UniqueConstraintModel]"""
 
@@ -764,8 +721,11 @@ class IntrospectAndStoreDatabaseCommand(MetaCommand):
         return constraints
 
     @classmethod
-    def introspect_indexes(cls, inspector, schema_name, relation_name) -> List[IndexModel]:
+    def introspect_indexes(
+        cls, inspector: InspectorProtocol, schema_name: str, relation_name: str
+    ) -> List[IndexModel]:
         """Introspect all indexes for a table or materialized view, describing the results as a List[IndexModel]"""
+
         indexes = []
 
         index_dicts = inspector.get_indexes(relation_name, schema_name)
@@ -783,48 +743,38 @@ class IntrospectAndStoreDatabaseCommand(MetaCommand):
 
     @classmethod
     def introspect_primary_key(
-        cls, inspector: SchemaStrippingInspector, relation_name: str, schema_name: str
+        cls, inspector: InspectorProtocol, relation_name: str, schema_name: str
     ) -> Tuple[Optional[str], List[str]]:
         """Introspect the primary key of a table, returning the pkey name and list of columns in the primary key (if any).
 
         If no primary key index is defined, will return None for name, empty list for columns.
         """
+
         primary_index_dict = inspector.get_pk_constraint(relation_name, schema_name)
 
-        # Athena dialect returns ... an empty _list_ instead of a dict, contrary to what
-        # https://docs.sqlalchemy.org/en/14/core/reflection.html#sqlalchemy.engine.reflection.Inspector.get_pk_constraint
-        # specifies for the return result from inspector.get_pk_constraint().
-        if isinstance(primary_index_dict, dict):
-            # MySQL at least can have unnamed primary keys. The returned dict will have 'name' -> None.
-            # Sigh.
-            pkey_name = primary_index_dict.get('name') or '(unnamed primary key)'
-
+        if primary_index_dict:
             if primary_index_dict['constrained_columns']:
-                return pkey_name, primary_index_dict['constrained_columns']
+                return primary_index_dict['name'], primary_index_dict['constrained_columns']
 
         # No primary key to be returned.
         return None, []
 
     @classmethod
     def introspect_columns(
-        cls, inspector: SchemaStrippingInspector, schema_name: str, relation_name: str
+        cls, inspector: InspectorProtocol, schema_name: str, relation_name: str
     ) -> List[ColumnModel]:
         column_dicts = inspector.get_columns(relation_name, schema=schema_name)
 
         retlist = []
 
         for col in column_dicts:
-            comment = col.get('comment')  # Some dialects do not return.
-
-            type_name = determine_column_type_name(col['type'])
-
             retlist.append(
                 ColumnModel(
                     name=col['name'],
                     is_nullable=col['nullable'],
-                    data_type=type_name,
+                    data_type=col['type'],
                     default_expression=col['default'],
-                    comment=comment,
+                    comment=col.get('comment'),  # Some dialects do not return this member at all.
                 )
             )
 
@@ -956,7 +906,7 @@ class RelationStructureMessager:
 
 
 def constraints_dataframe(
-    inspector: SchemaStrippingInspector, table_name: str, schema: Optional[str]
+    inspector: InspectorProtocol, table_name: str, schema: Optional[str]
 ) -> DataFrame:
     """Transform results from inspector.get_check_constraints() into a single dataframe for display() purposes"""
 
@@ -980,7 +930,7 @@ def constraints_dataframe(
 
 
 def foreignkeys_dataframe(
-    inspector: SchemaStrippingInspector, table_name: str, schema: Optional[str]
+    inspector: InspectorProtocol, table_name: str, schema: Optional[str]
 ) -> DataFrame:
     """Transform results from inspector.get_indexes() into a single dataframe for display() purposes"""
 
@@ -1016,7 +966,7 @@ def foreignkeys_dataframe(
 
 
 def index_dataframe(
-    inspector: SchemaStrippingInspector, table_name: str, schema: Optional[str]
+    inspector: InspectorProtocol, table_name: str, schema: Optional[str]
 ) -> DataFrame:
     """Transform results from inspector.get_indexes() into a single dataframe for display() purposes"""
 
@@ -1217,19 +1167,6 @@ def _raise_from_no_such_table(schema: str, relation_name: str):
     else:
         msg = f'Relation {relation_name} does not exist'
     raise MetaCommandException(msg)
-
-
-def determine_column_type_name(sqla_column_type_object: TypeEngine) -> str:
-    """Convert the possibly db-centric TypeEngine instance to a sqla-generic type string"""
-    try:
-        type_name = str(sqla_column_type_object.as_generic()).lower()
-    except (NotImplementedError, AssertionError):
-        # ENG-5268: More esoteric types like UUID do not implement .as_generic()
-        # ENG-5808: Some Databricks types are not fully implemented and fail
-        # assertions within .as_generic()
-        type_name = str(sqla_column_type_object).replace('()', '').lower()
-
-    return type_name
 
 
 def make_introspection_error_human_presentable(exception: Exception) -> str: